Make ExistsFn usable by introducing a Data type.
@Quanyails noticed its not very ergonomic of feasible to implement
an ExistsFn for Generations, but trying to make it usable revealed
several other shortcomings.

- Export a `Data` type. Logically having a `Data` type makes sense
given `DataKind` already exists, but `Data` is both a poor name and
can be confused with the `FooData`-type types for Data `Foo` (say
that 5 times fast). Somewhat regrettable name, but the type is
required to be able to export something for `ExistsFn`
- Make it so all kinds of `Data` are subject to the `ExistsFn`
instead of effects + species. This was mostly not done out of
laziness/lack of need before, but supporting all kinds is more
logical and consistent.
- Fix `@pkmn/sim`'s `Dex` not actually setting the `kind` field...
whoops.
- Fix `@pkmn/sim` always claiming `Learnsets` exist (wtf PS?)
- Add tests for `exists` override
- Move `Generations` cache inside `Generations` instead of a weird
global cache that broke tests (why was it ever written that way to
begin with???)